Few changes to movies to make it work

+++ liveIt-service/src/main/java/com/liveit/service/MovieService.java
@@ -12,7 +12,7 @@ public interface MovieService {
 	
 	List<MovieEntity> getMovies();
 	
-	MovieEntity getMovie(long id) throws Exception;
+	MovieEntity getMovie(int id) throws Exception;
 
 	MovieEntity update(MovieVo movie) throws ParseException;
